Move RuntimeRequest / Response to top level
