+++ spec/request_validation/middleware_spec.rb
@@ -26,7 +26,7 @@ RSpec.describe OpenapiFirst::RequestValidation::Middleware do
 
     it 'adds request to env ' do
       get '/pets'
-      expect(last_request.env[OpenapiFirst::REQUEST]).to be_a OpenapiFirst::Definition::RuntimeRequest
+      expect(last_request.env[OpenapiFirst::REQUEST]).to be_a OpenapiFirst::RuntimeRequest
     end
   end
